+++ src/gfx/sprite2d.h
@@ -7,7 +7,7 @@
 
 #define SPRITES_MAX 2000
 
-bool sprite2d_init(unsigned long viewport_width, unsigned long viewport_height);
+bool sprite2d_init(unsigned int viewport_width, unsigned int viewport_height);
 void sprite2d_shutdown();
 
 void sprite2d_begin();
